+++ app/controllers/platform/organizations_controller.rb
@@ -3,14 +3,17 @@
 module Platform
   class OrganizationsController < BaseController
     def index
+      authorize Organization
       @organizations = Organization.order(:name)
     end
 
     def new
+      authorize Organization
       @organization = Organization.new
     end
 
     def create
+      authorize Organization
       @organization = Organization.new(organization_params)
       if @organization.save
         redirect_to platform_organizations_path, notice: "Организация создана"
@@ -21,10 +24,12 @@ module Platform
 
     def edit
       @organization = Organization.find(params[:id])
+      authorize @organization
     end
 
     def update
       @organization = Organization.find(params[:id])
+      authorize @organization
       if @organization.update(organization_params)
         redirect_to platform_organizations_path, notice: "Сохранено"
       else